plat: Use PC to control protection (psoc64)
With older versions of Secure Flashboot (SFB), the CM0 is placed
into non-secure mode before TFM is run, and the registers are locked
down so that Cy_Prot_ConfigBusMaster() will cause a hard fault.

This change alters the SMPU setup to reflect that environment.

Protection Context (PC) is used to distinguish Secure from Non-secure,
while privileged/user is still used in the same manner.
We rely on SFB restricting the ability of the two cores to change their
PC.

Change-Id: I87d34a319f5c0b05831796ea3a113b3a54b61e1c
Signed-off-by: Chris Brand <chris.brand@cypress.com>