Build: Add ability to sign S and NS images separately (psoc64, psoc64_1m)
With this change, script can sign each image individually or
both in the same time.

As long as cysecuretools sign_image overwrites unsigned image file
with the signed one, script creates a copy of the original image
and signs the copy.

For CM4, sign_image also creates an unsigned copy of the image
named <image to sign>_cm4.hex. Script deletes it to avoid confusion.

Signed-off-by: Andrei Narkevitch <ainh@cypress.com>
Change-Id: I99381c33d68c2281cae2ac55cc482da7db5f78ac