Platform: Re-arrange regions in scatter file
This patch re-arranges regions in Armclang scatter file for the
following reason:
For stronger security, it is better to have only code is executable.
This means RO data should not be executable.
While currently code and RO data are mixed together, thus they would have
same attributes - read-only and executable.

This patch:
1. Splits CODE and RO-DATA to different regions so that they can have
   different attributes, but they are still neibour to each other so that
   CODE + RO-DATA together still can be one single region to save MPU
   regions.
2. Moves platform specific regions together to the bottom

Note:
This patch will slightly increase the memory footprint due to the need
to align the RO Data to the 32 byte memory.

Change-Id: Ifda64e060f5d9b68056deefe1b24c131237766ce
Signed-off-by: Kevin Peng <kevin.peng@arm.com>